+++ src/components/my-studio/review-list.tsx
@@ -1,4 +1,4 @@
-import { getMyReviews } from '@/service/client-actions/studio';
+import { getMyReviews } from '@/service/server-actions/studio';
 import { Textarea } from '../ui/textarea';
 import { cn } from '@/lib/utils';
 
@@ -9,11 +9,9 @@ interface IReviewListProps {
 const ReviewList = async ({ id }: IReviewListProps) => {
   const reviewList = await getMyReviews(id);
 
-  console.log(reviewList);
-
   return (
     <section>
-      {/* {reviewList.length <= 0 && (
+      {reviewList.length <= 0 && (
         <div className="mt-[48px] text-center">
           <div className="font-semibold leading-[135%] text-gray-20">
             아직 작성된 리뷰가 없습니다.
@@ -27,7 +25,7 @@ const ReviewList = async ({ id }: IReviewListProps) => {
       <ul className="space-y-4">
         {reviewList.map((review, i) => (
           <li
-            key={review.nickname}
+            key={review.reviewerNickname + i}
             className={cn(
               'pb-[16px]',
               i === reviewList.length - 1
@@ -36,10 +34,10 @@ const ReviewList = async ({ id }: IReviewListProps) => {
             )}
           >
             <div className="text-lg font-semibold leading-[135%]">
-              {review.nickname}
+              {review.reviewerNickname}
             </div>
             <div className="mt-2 flex flex-wrap items-center gap-[6px]">
-              {review.tagList.map((tag, i) => (
+              {review.tags.map((tag, i) => (
                 <span
                   key={i + tag}
                   className="inline-flex items-center justify-center rounded-[8px] bg-gray-90 p-2 text-xs text-gray-10"
@@ -48,16 +46,16 @@ const ReviewList = async ({ id }: IReviewListProps) => {
                 </span>
               ))}
             </div>
-            {review.constents && (
+            {review.content && (
               <Textarea
                 readOnly
                 className="mt-[10px] resize-none focus-visible:ring-0 focus-visible:ring-offset-0"
-                value={review.constents}
+                value={review.content}
               />
             )}
           </li>
         ))}
-      </ul> */}
+      </ul>
     </section>
   );
 };
