+++ src/service/server-actions/studio.ts
@@ -1,0 +1,36 @@
+import { cookies } from 'next/headers';
+
+const API_URL = process.env.NEXT_PUBLIC_API_URL;
+
+export const getMyReviews = async (id: number) => {
+  const cookieStore = cookies();
+  const token = cookieStore.get('accessToken');
+
+  if (token) {
+    try {
+      const res = await fetch(`${API_URL}/users/${id}/projects/reviews`, {
+        method: 'GET',
+        cache: 'no-store',
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${token?.value}`,
+        },
+      });
+
+      // if (!res.ok) {
+      //   throw new Error('something error : ' + res.status);
+      // }
+
+      const data: {
+        reviewerNickname: string;
+        tags: string[];
+        content: string;
+      }[] = await res.json();
+
+      return data;
+    } catch (error) {
+      console.log(error);
+    }
+  }
+  return [];
+};
